+++ src/builder/app.ts
@@ -1,4 +1,5 @@
 import { AppView } from "../types/app";
+import { trace } from "../reader/tracer";
 import { BuildContext } from "./context";
 
 export function buildApp(context: BuildContext, view: AppView) {
