+++ src/reader/modifiers/backgroundColor.ts
@@ -1,7 +1,7 @@
 import { FigmaContext } from "../context";
 import { Color, View } from "../../types/views";
 import { BackgroundModifier } from "../../types/modifiers";
-import { trace } from "../../util/tracer";
+import { trace } from "../tracer";
 
 export function appendBackgroundColor(
   context: FigmaContext,
