+++ src/reader/modifiers/position.ts
@@ -2,7 +2,7 @@ const assert = require("assert");
 import { FigmaContext } from "../context";
 import { PositionModifier } from "../../types/modifiers";
 import { isAxisView, View } from "../../types/views";
-import { trace } from "../../util/tracer";
+import { trace } from "../tracer";
 
 export function appendPosition(
   context: FigmaContext,
