+++ src/reader/node/componentNode.ts
@@ -1,5 +1,5 @@
 import { ZStack } from "../../types/views";
-import { trace } from "../../util/tracer";
+import { trace } from "../tracer";
 import { FigmaContext } from "../context";
 import { traverse } from "../entrypoint";
 import { appendBackgroundColor } from "../modifiers/backgroundColor";
