+++ src/reader/node/shapeWithTextNode.ts
@@ -1,4 +1,4 @@
-import { trace } from "../../util/tracer";
+import { trace } from "../tracer";
 import { FigmaContext } from "../context";
 
 export function walkToShapeWithText(
