+++ src/reader/view/image.ts
@@ -3,7 +3,7 @@ import { ImageModifier } from "../../types/imageModifier";
 import { Image } from "../../types/views";
 import { appendForegroundColor } from "../modifiers/foregroundColor";
 import { appendFixedFrame } from "../modifiers/frame";
-import { trace } from "../../util/tracer";
+import { trace } from "../tracer";
 
 export type ImageNode = DefaultShapeMixin & SceneNode;
